refactor: move navigation-related functions in separate module
