+++ app/sitemap.ts
@@ -6,7 +6,7 @@ import type { MetadataRoute } from "next";
 
 import { env } from "@/config/env.config";
 import { locales } from "@/lib/i18n/locales";
-import { getPathname } from "@/lib/i18n/navigation";
+import { getPathname } from "@/lib/navigation/navigation";
 
 const baseUrl = env.NEXT_PUBLIC_APP_BASE_URL;
