+++ components/link.tsx
@@ -12,7 +12,7 @@ import {
 } from "react-aria";
 import type { LinkProps as AriaLinkProps } from "react-aria-components";
 
-import { LocaleLink, type LocaleLinkProps } from "@/lib/i18n/navigation";
+import { LocaleLink, type LocaleLinkProps } from "@/lib/navigation/navigation";
 import { useRenderProps } from "@/lib/use-render-props";
 
 /**
