+++ eslint.config.ts
@@ -52,12 +52,18 @@ const config: Config = [
 				},
 				{
 					name: "next/navigation",
-					importNames: ["redirect", "permanentRedirect", "useRouter", "usePathname"],
-					message: "Please use `@/lib/i18n/navigation` instead.",
+					importNames: [
+						"permanentRedirect",
+						"redirect",
+						"usePathname",
+						"useRouter",
+						"useSearchParams",
+					],
+					message: "Please use `@/lib/navigation/navigation` instead.",
 				},
 				{
 					name: "next/router",
-					message: "Please use `@/lib/i18n/navigation` instead.",
+					message: "Please use `@/lib/navigation/navigation` instead.",
 				},
 			],
 			"no-restricted-syntax": [
