+++ lib/analytics-script.tsx
@@ -2,7 +2,6 @@
 
 import { createUrl } from "@acdh-oeaw/lib";
 import type { NextWebVitalsMetric } from "next/app";
-import { useSearchParams } from "next/navigation";
 import Script from "next/script";
 import { useReportWebVitals } from "next/web-vitals";
 import { useLocale } from "next-intl";
@@ -10,7 +9,7 @@ import { Fragment, type ReactNode, Suspense, useEffect } from "react";
 
 import { env } from "@/config/env.config";
 import type { IntlLocale } from "@/lib/i18n/locales";
-import { usePathname } from "@/lib/i18n/navigation";
+import { usePathname, useSearchParams } from "@/lib/navigation/navigation";
 
 declare global {
 	interface Window {
