+++ lib/navigation/navigation.ts
@@ -16,6 +16,9 @@ const redirect: typeof _redirect = _redirect;
 
 export { getPathname, redirect, usePathname, useRouter };
 
+// eslint-disable-next-line no-restricted-imports
+export { useSearchParams } from "next/navigation";
+
 export type LocaleLinkProps = Omit<ComponentPropsWithRef<typeof Link>, "href"> & {
 	href?: string | undefined;
 };
